// src/interfaces/hero.interface.ts
// Generated by https://quicktype.io

import { Stories, Thumbnail, URLType } from "./response.interface";

export interface Hero {
  id:          number;
  name:        string;
  description: string;
  modified:    string;
  thumbnail:   Thumbnail;
  resourceURI: string;
  comics:      Comics;
  series:      Comics;
  stories:     Stories;
  events:      Comics;
  urls:        URLType[];
}

export interface Comics {
  available:     number;
  collectionURI: string;
  items:         ComicsItem[];
  returned:      number;
}

export interface ComicsItem {
  resourceURI: string;
  name:        string;
}